+++ osu.Game/Overlays/BeatmapSet/BeatmapPicker.cs
@@ -87,7 +87,7 @@ namespace osu.Game.Overlays.BeatmapSet
                                 {
                                     Anchor = Anchor.BottomLeft,
                                     Origin = Anchor.BottomLeft,
-                                    Font = OsuFont.GetFont(size: 11, weight: FontWeight.Bold),
+                                    Font = OsuFont.GetFont(size: 15, weight: FontWeight.Bold),
                                     Text = "难度星级",
                                     Alpha = 0,
                                     Margin = new MarginPadding { Bottom = 1 },